implement top level postgres topic repo implementation (expect existing, find, search)

+++ engine/src/pagination.rs
@@ -1,13 +1,13 @@
 use serde::Deserialize;
 use utoipa::ToSchema;
 
-const fn default_page() -> usize {
+const fn default_page() -> u32 {
     1
 }
 
 #[derive(Debug, Deserialize, ToSchema, PartialEq, Eq)]
 pub struct Pagination {
     #[serde(default = "default_page")]
-    pub page: usize,
-    pub page_size: Option<usize>,
+    pub page: u32,
+    pub page_size: Option<u32>,
 }
